pyproject.toml
[tool.poetry]
name = "store-api"
version = "0.1.0"
description = ""
authors = ["ku2h7 <kukuhsyaifudin7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
poetry-dotenv-plugin = "^0.2.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
